se agrego RSSI y se arreglo bug firmware version

+++ lib/settings/settings.h
@@ -34,7 +34,7 @@ struct RTCStatus
 
 struct Settings
 {
-    uint16_t version;
+    float version;
 
     WifiSettings wifi;
